Refined Image sizing + training code
This is a refined version of the previous codes to train the model. It successfully saves and resizes the training images; however, it is taking quite a while to train (maybe its my laptop). If anyone else wants to try running this code, you just have to change the initial path to wherever your train dataset is saved on your laptop and run!